Make detecting errors easier
Apparently, CMake won't exit with a non-zero return code in script mode
unless the user calls cmake_exit().

To detect errors, we can only check whether the output we receive is
empty or malformed. By default, CMake won't write to stdout, but to stderr.
We can explicitly write to stdout and capture only stdout from the
Python side. If the string is empty, we can assume there was some kind
of issue.